[feat] 상세페이지 다시 연결_yr[220510]
