+++ src/components/search/index.js
@@ -1,0 +1,5 @@
+import SearchBar from './SearchBar';
+import SearchHistory from './SearchHistory'
+
+
+export { SearchBar, SearchHistory };
