+++ src/shared/App.js
@@ -14,7 +14,6 @@ import {
   SearchPage,
   NotFound,
   SearchList,
-  SearchMap,
   DetailOffice,
   MapOfficeList,
 } from "../pages/index";
@@ -45,7 +44,6 @@ function App() {
               <Route path="/like" exact component={Like} />
               <Route path="/searchlist" exact component={SearchList} />
               <Route path="/map/office" exact component={MapOfficeList} />
-              <Route path="/searchmap/:estateId" exact  component={SearchMap}/>
               <Route path="/detail/:estateId" exact  component={DetailOffice}/>
               <Route component={NotFound} />
             </Switch>
